// src/app/model/teacher.ts
export interface Teacher {
  teacherUserId: number;
  addressId: number;
  teacherUsername: string;
  teacherPassword: string;
  firstName: string;
  lastname: string;
  phoneNumber: string;
  email: string;
  dateOfBirth: string;
  subjectId: string;
  doorNum: string;
  street: string;
  addrLine: string;
  city: string;
  state: string;
  pincode: string;
  gender: string;
}